code adjustment js
