+++ BackEnd/src/main/java/at/technikumwien/webshop/config/SecurityConfig.java
@@ -56,7 +56,7 @@ public class SecurityConfig {
                 .requestMatchers(HttpMethod.PUT, "/users/update", "/products/update", "/files/update").hasRole("ADMIN")
                 .requestMatchers(HttpMethod.DELETE, "/users/{id}", "/products/{id}").hasRole("ADMIN")
                 //all user requests
-                .requestMatchers(HttpMethod.POST, "/users/createUser").permitAll()
+                .requestMatchers(HttpMethod.POST, "/users/createUser", "/positions").permitAll()
                 .requestMatchers(HttpMethod.GET, "/files/**", "/products/active").permitAll()
                 .requestMatchers("/login", "/addresses", "/addresses/users/**", "/users/{id}/address").permitAll()
