+++ BackEnd/src/main/java/at/technikumwien/webshop/model/User.java
@@ -2,7 +2,6 @@ package at.technikumwien.webshop.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
