+++ FrontEnd/components/navbar/overlay.html
@@ -23,7 +23,7 @@
           src="/frontend/components/navbar/img/jumping_frog_logo.png" width="100" height="50" alt="logo"></a>
       <div>
         <a id="adminInterfaceLink"></a>
-        <a href="/frontend/index.html"><img class="navbar-img mx-3 my-1"
+        <a href="/frontend/others.html"><img class="navbar-img mx-3 my-1"
             src="/frontend/components/navbar/img/sale_48.png" width="40" height="40" alt="sale"></a>
         <span id="userInterfaceLink"></span>
         <button id="loginButtonOverlay" type="button" class="btn btn-outline-secondary text-white ms-3"
