+++ FrontEnd/components/productCard/productCard.html
@@ -1,0 +1,71 @@
+<div class="card bg-dark border border-5 border-light text-white p-3" style="width: 22rem;">
+    <img class="card-img-top border border-1 border-light rounded" id="showImage" height="350" src="">
+    <div class="card-body border border-1 border-bottom-0 rounded-top-1 mt-1 ">
+        <h5 class="card-title text-center" id="showName"></h5>
+        <div class="d-flex justify-content-center mb-1 mt-2">
+            <button type="button" class="btn btn-light" id="productDetailsModalButton"
+                style="width: 1.5rem; height: 1.5rem; padding: 0;" data-bs-toggle="modal"
+                data-bs-target="#productDetailsModal" data-toggle="tooltip" data-placement="top"
+                title="Beschreibung anzeigen">
+                <p>📜</p>
+            </button>
+            <div class="modal modal-lg fade" id="productDetailsModal" tabindex="-1" aria-labelledby="myModalLabel"
+                aria-hidden="true">
+                <div class="modal-dialog">
+                    <div class="modal-content bg-dark text-white">
+                        <div class="modal-body">
+                            <div class="row">
+
+                                <div class="col-4">
+                                    <div class="d-flex align-items-start h-100">
+                                        <img id="showModalImage" src="" alt="" class="img-fluid border border-light">
+                                    </div>
+                                </div>
+                                <div class="col-8 d-flex flex-column">
+                                    <div>
+                                        <h3 class="modal-title" id="showModalName"></h3>
+                                    </div>
+                                    <div class="mt-3">
+                                        <p id="showModalDescription">
+
+                                        </p>
+                                    </div>
+                                    <div class="mt-auto">
+                                        <div class="d-flex justify-content-between align-items-center">
+                                            <div class="d-flex align-items-center">
+                                                <p class="card-text mb-0" id="showModalPrice"></p>
+                                            </div>
+                                            <div class="ms-auto">
+                                                <p class="card-text mb-0" id="showModalQuantity"></p>
+                                            </div>
+                                            <div class="ms-3">
+                                                <button type="button" class="btn btn-light d-flex addToCartButtonClass" id="addToCartButtonModal"
+                                                    style="width: 1.5rem; height: 1.5rem; padding: 0;" role="button"
+                                                    data-toggle="tooltip" data-placement="top"
+                                                    title="Produkt zum Warenkorb hinzufügen">
+                                                    <p style="pointer-events: none;">➕</p>
+                                                </button>
+                                            </div>
+                                        </div>
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+    <div class="card-footer border border-1 border-top-1 d-flex justify-content-between">
+        <p class="card-text mb-0" id="showPrice"></p>
+        <button type="button" class="btn btn-light d-flex addToCartButtonClass" id="addToCartButton"
+            style="width: 1.5rem; height: 1.5rem; padding: 0;" role="button" data-toggle="tooltip" data-placement="top"
+            title="Produkt zum Warenkorb hinzufügen">
+            <p style="pointer-events: none;">➕</p>
+        </button>
+    </div>
+</div>
+
+
+
+<!--Product Modal-->
