+++ FrontEnd/index.html
@@ -8,13 +8,6 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
-
-  <!--Eigenen Files-->
-  <script src="./scripts/index.js" defer></script>
-  <script src="./scripts/products.js" defer></script>
-  <script src="/frontend/components/manaSymbols/manaSymbols.js" defer></script>
-  <script src="/frontend/components/magicCurser/magicCurser.js" defer></script>
-
   <link rel="stylesheet" href="style.css">
 
   <!--Bootstrap CSS-->
@@ -33,11 +26,11 @@
   <!--Overlay-Header-->
   <script src="/frontend/components/navbar/loadOverlay.js" defer></script>
   <script src="/frontend/scripts/login.js" defer></script>
-  
-  
-  
 
-  
+    <!--Eigenen Files-->
+    <script src="./scripts/index.js" defer></script>
+    <script src="/frontend/components/manaSymbols/manaSymbols.js" defer></script>
+    <script src="/frontend/components/productCard/productCard.js" defer></script>
 
 </head>
