+++ FrontEnd/others.html
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <title>Impressum</title>
+
+  <meta charset="UTF-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+  <!--CSS-->
+  <link rel="stylesheet" href="style.css">
+
+  <!--Bootstrap CSS-->
+  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css">
+
+  <!-- Bootstrap JS-->
+  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js" defer></script>
+
+
+  <!--Overlay-Header-->
+  <script src="/frontend/components/navbar/loadOverlay.js" defer></script>
+  <script src="/frontend/scripts/login.js" defer></script>
+
+</head>
+
+<body class="bg-dark">
+  <overlay></overlay>
+
+
+  <footer>
+
+  </footer>
+
+</body>
+
+</html>
