--- FrontEnd/pages/adminInterface.html
@@ -1,125 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <title>AdminInterface</title>
-
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-
-    <!--Eigenen Files-->
-    <script src="/frontend/components/manaSymbols/manaSymbols.js" defer></script>
-    <script src="../scripts/createProduct.js" defer></script>
-    <script src="../scripts/adminInterfaceUser.js" defer></script>
-    <script src="../scripts/adminInterfaceProduct.js" defer></script>
-
-    <link rel="stylesheet" href="../style.css">
-
-    <!--Bootstrap CSS-->
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/css/bootstrap.min.css">
-    <!-- Bootstrap JS-->
-    <script defer src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js"></script>
-
-    <!--Jquery Datenbank-->
-    <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-
-    <!--Overlay-Header-->
-    <script src="/frontend/components/navbar/loadOverlay.js" defer></script>
-    <script src="/frontend/scripts/login.js" defer></script>
-
-
-
-</head>
-
-<body class="bg-dark">
-    <overlay></overlay>
-    <h1 class="p-3 text-center text-white">Admin Seite</h1>
-    <div class="container">
-        <div class="row">
-            <div class="col-10 col-lg-7 col-xxl-5 border border-dark rounded p-3 bg-light">
-                <div class="container">
-                    <h2 class="text-center mb-0">Produkt erstellen</h2>
-                    <form id="formCreateProduct" method="POST">
-                        <div class="row mb-3">
-                            <label for="inputProductName" class="form-label p-0">Titel</label>
-                            <input type="text" class="form-control" id="inputProductName" name="inputProductName">
-                        </div>
-                        <div class="row mb-3">
-                            <div class="col-6 col-sm-3 mb-2 p-0">
-                                <label for="inputProductPrice" class="form-label">Preis</label>
-                                <div class="input-group">
-                                    <input type="text" class="form-control" id="inputProductPrice"
-                                        name="inputProductprice" aria-label="Euro amount">
-                                    <span class="input-group-text">€</span>
-                                </div>
-                            </div>
-                            <div class="col-6 col-sm-3">
-                                <label for="inputProductQuantity" class="form-label">Menge</label>
-                                <input type="number" class="form-control" id="inputProductQuantity"
-                                    name="inputProductQuantity">
-                            </div>
-                            <div class="col p-0">
-                                <label for="inputProductImg" class="form-label">Bild hochladen</label>
-                                <input class="form-control" type="file" id="inputProductImg" name="inputProductImg"
-                                    >
-                            </div>
-                        </div>
-                        <div class="row mb-3">
-                            <label for="inputProductDescription" class="form-label p-0">Beschreibung</label>
-                            <textarea class="form-control" id="description" name="description" rows="3"
-                                required></textarea>
-                        </div>
-                        <div class="row mb-2">Mana</div>
-                        <div class="row justify-content-between mb-3">
-                            <manaSymbols></manaSymbols>
-                        </div>
-                        <div class="row justify-content-center">
-                            <button class="btn btn-success mt-3" type="submit" id="createProduct">&#x2795; Produkt erstellen</button>
-                        </div>
-                    </form>
-                </div>
-            </div>
-            <div class="col-xxl-7 pe-0">
-                <div class="container">
-                    <div class="row">
-                        <div class="col-12 border border-dark rounded p-3 bg-light">
-                            <div class="container" id="productListContainer">
-                                <h2 class="text-center mb-0">Prokdukt Liste</h2>
-                            </div>
-                            <div class="row text-center" id="pagesButton">
-                                <div class="col text-end"><button id="previousPage" class="btn btn-dark">Vorherige Seite</button></div>
-                                <div class="col-1 d-flex align-items-center justify-content-center"><p class="m-0"></p></div>
-                                <div class="col text-start"><button id="nextPage" class="btn btn-dark">Nächste Seite</button></div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-    <div class="container mt-3">
-        <div class="row">
-            <div class="col-12 border border-dark rounded p-3 bg-light">
-                <div class="container">
-                    <h1 class="text-center">Benutzer Liste</h1>
-                    <table class="table table-striped">
-                        <thead>
-                            <tr>
-                                <th>ID</th>
-                                <th>Benutzername</th>
-                                <th>E-Mail</th>
-                                <th>Aktiv</th>
-                                <th>Admin</th>
-                            </tr>
-                        </thead>
-                        <tbody id="userTableBody"></tbody>
-                    </table>
-                </div>
-            </div>
-        </div>
-    </div>
-
-</body>
-
-</html>
