+++ FrontEnd/scripts/adminInterfaceUser.js
@@ -1,32 +1,18 @@
+//Benutzer als Liste hinzufügen
+
+$(document).ready(function () {
+    loadUsers();
+});
+
 function loadUsers() {
     $.ajax({
         url: "http://localhost:8080/users",
         method: "GET",
         headers: { "Authorization": sessionStorage.getItem("token") },
         success: function (users) {
-            var userTableBody = $("#userTableBody");
-            userTableBody.empty();
-
-            for (var i = 0; i < users.length; i++) {
-                var user = users[i];
-                var row = $("<tr></tr>");
-                row.append($("<td class='align-middle'>" + user.id + "</td>"));
-                row.append($("<td class='align-middle'>" + user.username + "</td>"));
-                row.append($("<td class='align-middle'>" + user.email + "</td>"));
-                row.append($("<td class='align-middle'>" + (user.active ? "&#10004;&#65039;" : "&#10060;") + "</td>"));
-                row.append($("<td class='align-middle'>" + (user.admin ? "&#10004;&#65039;" : "&#10060;") + "</td>"));
-
-                var editButton = $("<button class='btn btn-primary'>	&#x2692;&#xFE0F; Bearbeiten</button>");
-                editButton.click(createEditUserHandler(user));
-
-                var deleteButton = $("<button class='btn btn-danger mx-1'>&#x1F5D1;&#xFE0F; Löschen</button>");
-                deleteButton.click(createDeleteUserHandler(user.id));
-
-                var buttonCell = $("<td class='text-end'></td>").append(editButton, deleteButton);
-                row.append(buttonCell);
-
-                userTableBody.append(row);
-            }
+            allUsers = users;
+            currentPage = 1;
+            addUsersToList(users);
         },
         error: function (error) {
             console.error(error);
@@ -34,126 +20,124 @@ function loadUsers() {
     });
 }
 
-function createEditUserHandler(user) {
-    return function () {
-        var userTableBody = $("#userTableBody");
-        userTableBody.empty();
+//Benutzer laden
 
-        var row = $("<tr></tr>");
+function addUsersToList(users) {
+    const divPages = $("#usersPagesButton");
+    divPages.show();
+    divPages.find("p").text(currentPage);
 
-        var idCell = $("<td class='align-middle'></td>").text(user.id);
-        row.append(idCell);
+    const usersPerPage = 6;
+    const startIndex = (currentPage - 1) * usersPerPage;
+    const endIndex = startIndex + usersPerPage;
+    const usersToShow = users.slice(startIndex, endIndex);
+    const userTableBody = $("#userTableBody");
 
-        var usernameInput = $("<input id='username' type='text' class='form-control' value='" + user.username + "'>");
-        var usernameCell = $("<td></td>").append(usernameInput);
-        row.append(usernameCell);
+    userTableBody.empty();
 
-        var emailInput = $("<input id='email' type='email' class='form-control' value='" + user.email + "'>");
-        var emailCell = $("<td></td>").append(emailInput);
-        row.append(emailCell);
+    displayUserEditOrList("table-row-group", "none");
 
-        var activeDropdown = $("<select class='form-control' id='activeDropdown'></select>");
-        activeDropdown.append($("<option value='true'>&#10004;&#65039;</option>"));
-        activeDropdown.append($("<option value='false'>&#10060;</option>"));
-        activeDropdown.val(user.active.toString());
-        var activeCell = $("<td></td>").append(activeDropdown);
-        row.append(activeCell);
-
-        var adminDropdown = $("<select class='form-control' id='adminDropdown'></select>");
-        adminDropdown.append($("<option value='true'>&#10004;&#65039;</option>"));
-        adminDropdown.append($("<option value='false'>&#10060;</option>"));
-        adminDropdown.val(user.admin.toString());
-        var adminCell = $("<td></td>").append(adminDropdown);
-        row.append(adminCell);
-
-        var saveButton = $("<button class='btn btn-primary mx-1'>&#x1F4BE; Speichern</button>");
-        saveButton.click(createSaveUserHandler(user));
-        var cancelButton = $("<button class='btn btn-secondary'>&#x2716;&#xFE0F; Abbrechen</button>");
-        cancelButton.click(function () {
-            loadUsers();
+    usersToShow.forEach(function (user) {
+        const row = $("<tr>");
+        row.append($("<td>").text(user.id));
+        row.append($("<td>").text(user.username));
+        row.append($("<td class='d-none d-md-table-cell'>").text(user.email));
+        row.append($("<td class='text-center'>").html(user.active ? "&#10004;&#65039;" : "&#10060;"));
+        row.append($("<td class='text-center'>").html(user.admin ? "&#10004;&#65039;" : "&#10060;"));
+        const userEditButton = $("<button class='btn btn-primary mx-1' id='userEditButton'><span class='text-nowrap'><span class='d-none d-sm-inline'>Bearbeiten </span><span>&#x2692;&#xFE0F;</span></span></button>");
+        userEditButton.click(function () {
+            editUser(user);
         });
+        const userDeleteButton = $("<button class='btn btn-danger mx-1' id='userDeleteButton'>&#x1F5D1;&#xFE0F;</button>");
+        userDeleteButton.click(function () {
+            deleteUser(user.id);
+        })
 
-        var buttonCell = $("<td class='text-end'></td>").append(saveButton, cancelButton);
-        row.append(buttonCell);
+        const userHandler = $("<td class='text-end pe-0'>");
+        userHandler.append(userEditButton, userDeleteButton);
+        row.append(userHandler);
 
         userTableBody.append(row);
-    };
+    });
 }
 
+function editUser(user) {
 
-function createSaveUserHandler(user) {
-    return function () {
-        var userId = user.id;
-        var newUsername = $("#username").val();
-        var newEmail = $("#email").val();
-        var isActive = $("#activeDropdown").val() === "true";
-        var isAdmin = $("#adminDropdown").val() === "true";
+    displayUserEditOrList("none", "table-row-group");
 
-        // Überprüfe, ob das Eingabefeld leer ist
-        if (newUsername.trim() === "") {
-            // Wenn das Eingabefeld leer ist, behalte den alten Benutzernamen bei
-            newUsername = user.username;
-        }
+    $("#editUserId").text(user.id);
+    $("#editUserName").val(user.username);
+    $("#editUserEmail").val(user.email);
 
-        if (newEmail.trim() === "") {
-            // Wenn das Eingabefeld leer ist, behalte die alte E-Mail-Adresse bei
-            newEmail = user.email;
-        }
+    $("#editUserActive").val(user.active.toString());
+    $("#editUserAdmin").val(user.admin.toString());
 
-        var updatedUser = {
-            id: userId,
-            username: newUsername,
-            email: newEmail,
-            password: user.password,
-            active: isActive,
-            admin: isAdmin,
-            address: user.address || {} // Füge eine leere Adresse hinzu, wenn keine Adresse vorhanden ist
-        };
+    $("#saveEditUser").off("click").on("click", function () {
+        saveUser(user);
+    });
 
-        console.log("updatedUser:", updatedUser);
+}
 
-        // AJAX PUT-Anfrage senden, um den Benutzer zu aktualisieren
+$(document).on('click', '#cancelEditUser', function () {
+    displayUserEditOrList("table-row-group", "none");
+});
+
+
+// Produkt löschen
+
+function deleteUser(userId) {
+    if (confirm("Are you sure you want to delete this user?")) {
         $.ajax({
-            url: "http://localhost:8080/users/update",
-            method: "PUT",
+            url: "http://localhost:8080/users/" + userId,
+            method: "DELETE",
             headers: { "Authorization": sessionStorage.getItem("token") },
-            data: JSON.stringify(updatedUser),
-            contentType: "application/json",
             success: function (response) {
-                // Hier kannst du den Code zum Handhaben der erfolgreichen Aktualisierung implementieren
-                console.log("Updated user:", response);
-
-                // Lade die Benutzer erneut, um die aktualisierten Daten anzuzeigen
+                console.log("Deleted user:", response);
                 loadUsers();
             },
             error: function (error) {
                 console.error(error);
             }
         });
+    }
+};
+
+
+function saveUser(user) {
+
+    // Aktualisierte Userinformationen
+    const updatedUser = {
+        id: user.id,
+        password: user.password,
+        username: $("#editUserName").val().trim() || user.name,
+        email: $("#editUserEmail").val().trim() || user.email,
+        active: $("#editUserActive").val() === "true",
+        admin: $("#editUserAdmin").val() === "true"
     };
-}
 
-
-
-function createDeleteUserHandler(userId) {
-    return function () {
-        if (confirm("Are you sure you want to delete this user?")) {
-            $.ajax({
-                url: "http://localhost:8080/users/" + userId,
-                method: "DELETE",
-                headers: { "Authorization": sessionStorage.getItem("token") },
-                success: function (response) {
-                    console.log("Deleted user:", response);
-                    loadUsers();
-                },
-                error: function (error) {
-                    console.error(error);
-                }
-            });
+    // User aktualisieren
+    $.ajax({
+        url: "http://localhost:8080/users/update",
+        method: "PUT",
+        headers: { "Authorization": sessionStorage.getItem("token") },
+        data: JSON.stringify(updatedUser),
+        contentType: "application/json",
+        success: function (response) {
+            console.log("Updated user: ", response);
+            loadUsers();
+        },
+        error: function (error) {
+            console.error(error);
         }
-    };
+    });
 }
 
-$(document).ready(function () {
-    loadUsers();
-});
+//Switch between List and Edit
+
+function displayUserEditOrList(displayUserList, displayUserEdit) {
+    let userListTable = document.getElementById("userTableBody");
+    let editUserTable = document.getElementById("editUserTable");
+
+    userListTable.style.display = displayUserList;
+    editUserTable.style.display = displayUserEdit;
+}
